+++ integration/lang_test.go
@@ -459,3 +459,7 @@ func TestExecStrIndex(t *testing.T) {
 func TestExecStrSplitLen(t *testing.T) {
 	assertOutput(t, `print(len("a b c".split()))`, "3\n")
 }
+
+func TestExecStrEq(t *testing.T) {
+	assertOutput(t, `print("abc" == "abd")`, "0\n")
+}